+++ lettre/src/lib.rs
@@ -48,11 +48,63 @@ pub use smtp::{ClientSecurity, SmtpTransport};
 pub use smtp::client::net::ClientTlsParameters;
 use std::fmt::{self, Display, Formatter};
 use std::io::Read;
+use std::error::Error as StdError;
+use std::str::FromStr;
+
+/// Error type for email content
+#[derive(Debug)]
+pub enum Error {
+    /// Missing from in envelope
+    MissingFrom,
+    /// Missing to in envelope
+    MissingTo,
+    /// Invalid email
+    InvalidEmailAddress,
+}
+
+impl StdError for Error {
+    fn description(&self) -> &str {
+        match *self {
+            Error::MissingFrom => "missing source address, invalid envelope",
+            Error::MissingTo => "missing destination address, invalid envelope",
+            Error::InvalidEmailAddress => "invalid email address",
+        }
+    }
+
+    fn cause(&self) -> Option<&StdError> {
+        None
+    }
+}
+
+impl Display for Error {
+    fn fmt(&self, fmt: &mut Formatter) -> Result<(), fmt::Error> {
+        fmt.write_str(self.description())
+    }
+}
+
+/// Email result type
+pub type EmailResult<T> = Result<T, Error>;
 
 /// Email address
 #[derive(PartialEq, Eq, Clone, Debug)]
 #[cfg_attr(feature = "serde-impls", derive(Serialize, Deserialize))]
-pub struct EmailAddress(pub String);
+pub struct EmailAddress(String);
+
+impl EmailAddress {
+    /// Creates a new `EmailAddress`. For now it makes no validation.
+    pub fn new(address: String) -> EmailResult<EmailAddress> {
+        // TODO make some basic sanity checks
+        Ok(EmailAddress(address))
+    }
+}
+
+impl FromStr for EmailAddress {
+    type Err = Error;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        EmailAddress::new(s.to_string())
+    }
+}
 
 impl Display for EmailAddress {
     fn fmt(&self, f: &mut Formatter) -> fmt::Result {
@@ -60,19 +112,98 @@ impl Display for EmailAddress {
     }
 }
 
-impl EmailAddress {
-    /// Creates a new email address
-    pub fn new(address: String) -> EmailAddress {
-        EmailAddress(address)
+/// Simple email envelope representation
+///
+/// We only accept mailboxes, and do not support source routes (as per RFC).
+#[derive(PartialEq, Eq, Clone, Debug)]
+#[cfg_attr(feature = "serde-impls", derive(Serialize, Deserialize))]
+pub struct Envelope {
+    /// The envelope recipients' addresses
+    ///
+    /// This can not be empty.
+    forward_path: Vec<EmailAddress>,
+    /// The envelope sender address
+    reverse_path: Option<EmailAddress>,
+}
+
+impl Envelope {
+    /// Creates a new envelope, which may fail if `to` is empty.
+    pub fn new(from: Option<EmailAddress>, to: Vec<EmailAddress>) -> EmailResult<Envelope> {
+        if to.is_empty() {
+            return Err(Error::MissingTo);
+        }
+        Ok(Envelope {
+            forward_path: to,
+            reverse_path: from,
+        })
+    }
+
+    /// Destination addresses of the envelope
+    pub fn to(&self) -> &[EmailAddress] {
+        self.forward_path.as_slice()
+    }
+
+    /// Source address of the envelope
+    pub fn from(&self) -> Option<&EmailAddress> {
+        self.reverse_path.as_ref()
+    }
+
+    /// Creates a new builder
+    pub fn builder() -> EnvelopeBuilder {
+        EnvelopeBuilder::new()
+    }
+}
+
+/// Simple email envelope representation
+#[derive(PartialEq, Eq, Clone, Debug, Default)]
+pub struct EnvelopeBuilder {
+    /// The envelope recipients' addresses
+    to: Vec<EmailAddress>,
+    /// The envelope sender address
+    from: Option<EmailAddress>,
+}
+
+impl EnvelopeBuilder {
+    /// Constructs an envelope with no recipients and an empty sender
+    pub fn new() -> Self {
+        EnvelopeBuilder {
+            to: vec![],
+            from: None,
+        }
+    }
+
+    /// Adds a recipient
+    pub fn to<S: Into<EmailAddress>>(mut self, address: S) -> Self {
+        self.add_to(address);
+        self
+    }
+
+    /// Adds a recipient
+    pub fn add_to<S: Into<EmailAddress>>(&mut self, address: S) {
+        self.to.push(address.into());
+    }
+
+    /// Sets the sender
+    pub fn from<S: Into<EmailAddress>>(mut self, address: S) -> Self {
+        self.set_from(address);
+        self
+    }
+
+    /// Sets the sender
+    pub fn set_from<S: Into<EmailAddress>>(&mut self, address: S) {
+        self.from = Some(address.into());
+    }
+
+    /// Build the envelope
+    pub fn build(self) -> EmailResult<Envelope> {
+        Envelope::new(self.from, self.to)
     }
 }
 
 /// Email sendable by an SMTP client
 pub trait SendableEmail<'a, T: Read + 'a> {
-    /// To
-    fn to(&self) -> Vec<EmailAddress>;
-    /// From
-    fn from(&self) -> EmailAddress;
+    /// Envelope
+    fn envelope(&self) -> Envelope;
     /// Message ID, used for logging
     fn message_id(&self) -> String;
     /// Message content
@@ -89,10 +220,8 @@ pub trait EmailTransport<'a, U: Read + 'a, V> {
 #[derive(Debug, Clone)]
 #[cfg_attr(feature = "serde-impls", derive(Serialize, Deserialize))]
 pub struct SimpleSendableEmail {
-    /// To
-    to: Vec<EmailAddress>,
-    /// From
-    from: EmailAddress,
+    /// Envelope
+    envelope: Envelope,
     /// Message ID
     message_id: String,
     /// Message content
@@ -102,14 +231,30 @@ pub struct SimpleSendableEmail {
 impl SimpleSendableEmail {
     /// Returns a new email
     pub fn new(
-        from_address: EmailAddress,
-        to_addresses: Vec<EmailAddress>,
+        from_address: String,
+        to_addresses: &[String],
+        message_id: String,
+        message: String,
+    ) -> EmailResult<SimpleSendableEmail> {
+        let to: Result<Vec<EmailAddress>, Error> = to_addresses
+            .iter()
+            .map(|x| EmailAddress::new(x.clone()))
+            .collect();
+        Ok(SimpleSendableEmail::new_with_envelope(
+            Envelope::new(Some(EmailAddress::new(from_address)?), to?)?,
+            message_id,
+            message,
+        ))
+    }
+
+    /// Returns a new email from a valid envelope
+    pub fn new_with_envelope(
+        envelope: Envelope,
         message_id: String,
         message: String,
     ) -> SimpleSendableEmail {
         SimpleSendableEmail {
-            from: from_address,
-            to: to_addresses,
+            envelope,
             message_id,
             message: message.into_bytes(),
         }
@@ -117,12 +262,8 @@ impl SimpleSendableEmail {
 }
 
 impl<'a> SendableEmail<'a, &'a [u8]> for SimpleSendableEmail {
-    fn to(&self) -> Vec<EmailAddress> {
-        self.to.clone()
-    }
-
-    fn from(&self) -> EmailAddress {
-        self.from.clone()
+    fn envelope(&self) -> Envelope {
+        self.envelope.clone()
     }
 
     fn message_id(&self) -> String {